add tryCatch decorator
