+++ src/common/decorators/index.ts
@@ -1,0 +1,3 @@
+import TryCatch from './try-catch';
+
+export {TryCatch};
