+++ src/modules/avatars/avatars.controller.ts
@@ -1,32 +1,19 @@
 import { Response, NextFunction } from 'express';
 import {avatarService} from './avatars.service';
 import { AuthRequest } from '../../common/types/types';
+import { TryCatch } from '../../common/decorators/';
 
-function TryCatch(target: any, propName: string, descriptor: PropertyDescriptor) {
-    const fn = descriptor.value;
-    descriptor.value = async (...args) => {
-        try {
-            await fn.apply(this, args);
-        } catch (e) {
-            const [, , next] = args;
-            next(e);
-        }
-    };
-}
 
 class AvatarsController {
+    @TryCatch
     public async setOne(req: AuthRequest, res: Response, next: NextFunction): Promise<void> {
-        try {
-            const avatar = await avatarService.setOne(
-                req.body.img,
-                req.body.format,
-                req.params.purpose,
-                req.body.groupId,
-                req.user);
-            res.send(avatar);
-        } catch (e) {
-            next(e);
-        }
+        const avatar = await avatarService.setOne(
+            req.body.img,
+            req.body.format,
+            req.params.purpose,
+            req.body.groupId,
+            req.user);
+        res.send(avatar);
     }
 }
